Refactor mysql and sqlite3 code base
